Fix error: conversion to 'std::vector<double>::size_type' {aka 'long unsigned int'} from 'int' may change the sign of the result [-Werror=sign-conversion]

+++ link/.local/include/load_avg.hpp
@@ -34,14 +34,14 @@ get_load_avg(int num_samples = max_samples)
 	else if (num_samples > max_samples)
 		num_samples = max_samples;
 
-	std::vector<double> load_avg(num_samples);
+	std::vector<double> load_avg(static_cast<size_t>(num_samples));
 
 	const int num_samples_ret = getloadavg(load_avg.data(), num_samples);
 
 	if (num_samples_ret == -1)
 		load_avg.clear();
 	else if (num_samples_ret != num_samples)
-		load_avg.resize(num_samples_ret);
+		load_avg.resize(static_cast<size_t>(num_samples_ret));
 
 	return load_avg;
 }
